web_app/requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
stripe
python-dotenv
Flask-Mail
python-slugify
